+++ pnpm/src/cmd/index.ts
@@ -26,8 +26,8 @@ import { parseCliArgs } from '../parseCliArgs.js'
 import { shorthands as universalShorthands } from '../shorthands.js'
 import type { PnpmOptions } from '../types.js'
 import * as bin from './bin.js'
-import * as ci from './ci.js'
 import * as clean from './clean.js'
+import * as ci from './cleanInstall.js'
 import { createHelp } from './help.js'
 import * as installTest from './installTest.js'
 import * as recursive from './recursive.js'
